requirements.txt
fastapi
uvicorn
pymongo
alembic
psycopg2-binary
slowapi
python-dotenv
PyJWT
passlib
# bcrpyt==4.0.1
python-multipart
email-validator